chatbot/indexer/requirements.txt
pypdf>=4.0.0
pdfplumber>=0.11.0
google-genai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
